requirements.txt
aiogram==3.0.0b6
peewee==3.15.4
python-dotenv==0.21.0